Use array_values to renumber arrays after removing some values
Further tweaking of step 2 of adding the custom user fields
field_family_name field_given_name, and field_member_of to the form
and the view. On uninstall, renumber the indexes of the
'dependencies.config' configuration to start from 0, so that on
exporting to YAML, we don't see the indexes!