Define filters to get emails for a comment's parent and grandparent
Sigh, if a comment doesn't have a parent and/or a grandparent, we get
non-expansion of tokens when trying to get an email address.

Add a workaround: define filters to get the email address of a
comment's parent and grandparent comment. If there is no
parent/grandparent, return an empty string.