Don't send notifications for new/updated org and register nodes
Don't send notifications after saving a new organisation or register
node, or when their metadata is updated.

Why? Because creation/updating of such nodes is currently restricted
to the control body, so they (probably?) don't need notification of
updates.

Hmm, but for now we do leave notifications on for updates to the item
status field.