Limit reviewer values; show as a select list
Require register item reviewers to be members of either the
administrator or control body roles.  (Existing values of the field
are allowed to continue. So, e.g., if someone stops being a member of
the control body role, a register item can still be updated without
removing that user from the reviewer field.)

On all of the entry forms, show the field_reviewer field as a select
list.