Add coi module as a prerequisite and enable it
Add the coi module as a prerequisite, and enable it. This has the
effect of indicating configuration overrides (i.e., made in
settings.php) within the administration user interface. Configuration
values that have been overridden are shown with their "real" value,
and editing the value in the form is disabled.