Add updated patch to structure_sync to support URL aliases
See the previous commit.

The off-the-shelf patch to structure_sync module doesn't work in force
mode, which we need to use in order to preserve tids.

So patch the patch so that it supports force mode.

Also, preserve tids, which we weren't doing anyway (oops).
(See https://www.drupal.org/project/structure_sync/issues/3032480 .)