Update patch set for rules module
One of the patches we've been using for the rules module has now been
incorporated into the dev branch. So now we will update the installer
to get that newer commit. We no longer use the "upcast" patch, and we
update the one remaining patch to a later version that applies cleanly
to the more recent commit.